+++ service.sh
@@ -1,2 +1,0 @@
-#!/system/bin/sh
-setprop is_expressive_design_enabled true